Moved the Products Repository class into a separate file.
